Progressive seek fix (#58)
Summary: Pause the pipeline if seeking in the playing state to synchronise the client pipeline and the server. gst_element_lost_state is called when seeking, this moves the pipeline to the paused state and waits for preroll. After preroll, the pipeline will re-initiate the playing state.
Type: Bugfix
Test Plan: YtCert Tests
Jira: RIALTO-262